+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changes to the extension.
 
 - "Indexing Twine Project..." status bar item no longer hides the "Run Twine Game" item when the former appears.
 - Chapbook: Fixed bug where renaming variables and properties didn't rename all variables or properties.
+- Chapbook: Completions for modifiers after a semicolon no longer have an extra space.
 
 ## [1.1.0] - 2024 12 01